Completed Task #5 and #6
To check if opencv is working on local machine, a test image was read and displayed. This is just to set base for openCV and make sure it works. Modifications will be done when the sprint for iteration 1 begins. This addition is just for reference and a new development branch will be created for the sprint. Build issue on travis will be resolved first before branch is created.