Completed Task #11, #12, #13. Added functionalities.
Added the functionality to extract a region of interest (ROI) from an image; To take a perspective transform of an image, i.e. warp it; To detect straight lines in am image using Houghlines algorithm. Purpose is to make sure these functionalities work, they will be modified if need be when the complete pipeline in made functional.