Completed Task #8, #14 and #15. Testing functionalities on a video source.
Task #8 was to read frames from a video source. Added functionalities for creating a masked image (it will be considered as an alternative to extracting a region ofinterest and based on final results either of these would be deployed) and for sorting lines by selecting valid lines (based on threshold) and sorting into right and left lanes respectively.